Testing error handling. Code is currently performed within the calculate function but can be prevented from even getting this far if we move it to the input we are doubling up as the calculators display.
Potentially inefficient to check on every keypress but would mean we can prevent ++ or +/ or -- etc. Unsure how to check combinations of invalid operators, especially preventing using regex!

This is a point at which a unit test can be written to determine combinations of mock input values entered to input.